requirements.txt
Django==2.1.1
djangorestframework==3.8.2
mysqlclient==1.3.13
pytz==2018.5
